refactor: move CLI classes to Cli namespace (closes #71)
