+++ tests/Unit/ApplicationTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Unit;
 
-use Igancev\WorkReporter\Application;
+use Igancev\WorkReporter\Cli\Application;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Input\ArgvInput;
