+++ tests/Unit/InitCommandTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Unit;
 
-use Igancev\WorkReporter\InitCommand;
+use Igancev\WorkReporter\Cli\InitCommand;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Command\Command;
