+++ tests/Unit/WorkReportCommandTest.php
@@ -7,6 +7,7 @@ namespace Tests\Unit;
 use ArrayIterator;
 use DateTimeImmutable;
 use Exception;
+use Igancev\WorkReporter\Cli\WorkReportCommand;
 use Igancev\WorkReporter\Config\Config;
 use Igancev\WorkReporter\Config\ConfigException;
 use Igancev\WorkReporter\Config\ConfigProvider;
@@ -24,7 +25,6 @@ use Igancev\WorkReporter\Source\SourceType;
 use Igancev\WorkReporter\Source\TimeEntriesSource;
 use Igancev\WorkReporter\Source\TimeEntriesSourceFactory;
 use Igancev\WorkReporter\TimeEntry;
-use Igancev\WorkReporter\WorkReportCommand;
 use Laravel\Prompts\ConfirmPrompt;
 use Laravel\Prompts\Prompt;
 use PHPUnit\Framework\Attributes\AllowMockObjectsWithoutExpectations;
